More documentation
- Documented more files and classes.
- Moved RawSerializeVisitor.h from public include folder to the klvp src folder.
- Add a Index.md file for the document main page.
