+++ klvp/README.md
@@ -1,0 +1,16 @@
+klvp Library
+=========
+
+# KLV Parser (klvp)
+
+The Key-Length-Value (KLV) Paresr (**klvp**) library provides classes to allow client applications to parse and decode
+STANAG 4609 metadata that conforms to the following standards: 
+
+- MISB ST 0601, UAS Datalink Local Set 
+- MISB ST 0102, Security Metadata Universal and Local Sets for Motion Imagery Data
+
+## Project Structure
+
+Project creates a static library called **klvp**. 
+
+
