+++ src/TestKLVSecuritySetParser.h
@@ -3,19 +3,21 @@
 
 #include "KLVSecuritySetPrintVisitor.h"
 
-class TestKLVSecuritySetParser :
-	public lcss::KLVSecuritySetParser
+namespace klv2xml
 {
-public:
-	TestKLVSecuritySetParser(void);
-	virtual ~TestKLVSecuritySetParser(void);
+	class TestKLVSecuritySetParser :
+		public lcss::KLVSecuritySetParser
+	{
+	public:
+		TestKLVSecuritySetParser(void);
+		virtual ~TestKLVSecuritySetParser(void);
 
-	virtual void onBeginSet(int len, lcss::TYPE type);
-	virtual void onElement( lcss::KLVElement& klv);
-	virtual void onEndSet();
-	virtual void onError(const char* errmsg, int pos);
+		virtual void onBeginSet(int len, lcss::TYPE type);
+		virtual void onElement(lcss::KLVElement& klv);
+		virtual void onEndSet();
+		virtual void onError(const char* errmsg, int pos);
 
-private:
-	KLVSecuritySetPrintVisitor	printVisitor_;
-};
-
+	private:
+		KLVSecuritySetPrintVisitor	printVisitor_;
+	};
+}
